+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/ProjectsController/TaskDataControllerTests.cs
@@ -134,5 +134,78 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             Assert.NotNull(result);
             Assert.Equal(task.Id.Value, result.Id!.Value);
         }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateSupplementalFundingAgreementTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateSupplementalFundingAgreementTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Conversion;
+            command.Cleared = true;
+            command.Saved = false;
+            command.Received = true;
+            command.Signed = true;
+            command.SignedSecretaryState = true;
+
+            // Act
+            await tasksDataClient.UpdateSupplementalFundingAgreementTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.True(existingTaskData.SupplementalFundingAgreementCleared);
+            Assert.True(existingTaskData.SupplementalFundingAgreementReceived);
+            Assert.True(existingTaskData.SupplementalFundingAgreementSaved);
+            Assert.True(existingTaskData.SupplementalFundingAgreementSent);
+            Assert.True(existingTaskData.SupplementalFundingAgreementSigned);
+            Assert.True(existingTaskData.SupplementalFundingAgreementSignedSecretaryState);
+        }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateSupplementalFundingAgreementTaskAsync_ShouldUpdate_TransferTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateSupplementalFundingAgreementTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<TransferTasksData>();
+            dbContext.TransferTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Transfer;
+            command.Cleared = true;
+            command.Saved = false;
+            command.Received = true;
+
+            // Act
+            await tasksDataClient.UpdateSupplementalFundingAgreementTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.TransferTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.True(existingTaskData.SupplementalFundingAgreementCleared);
+            Assert.True(existingTaskData.SupplementalFundingAgreementReceived);
+            Assert.True(existingTaskData.SupplementalFundingAgreementSaved); 
+        }
     }
 }