+++ src/Core/Dfe.Complete.Application/Mappers/AutoMapping.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Dfe.AcademiesApi.Client.Contracts;
+using Dfe.Complete.Application.Contacts.Models;
 using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Domain.Entities;
 
@@ -86,7 +87,9 @@ namespace Dfe.Complete.Application.Mappers
 				.ForMember(dest => dest.Census, opt => opt.Ignore())
 				.ForMember(dest => dest.MisEstablishment, opt => opt.Ignore());
 			CreateMap<TransferTasksData, TransferTaskDataDto>();
-			CreateMap<SignificantDateHistory, SignificantDateHistoryDto>();
+            CreateMap<ConversionTasksData, ConversionTaskDataDto>();
+			CreateMap<KeyContact, KeyContactDto>();
+            CreateMap<SignificantDateHistory, SignificantDateHistoryDto>();
 			CreateMap<SignificantDateHistoryReason, SignificantDateHistoryReasonDto>();
 		}
 	}