+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Customizations/CustomWebApplicationDbContextFactoryCustomization.cs
@@ -112,7 +112,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
 
                 services.AddCompleteApiClient<IProjectsClient, ProjectsClient>(config, client);
                 services.AddCompleteApiClient<IProjectGroupClient, ProjectGroupClient>(config, client);
-                services.AddCompleteApiClient<ITransferTasksDataClient, TransferTasksDataClient>(config, client);
+                services.AddCompleteApiClient<ITasksDataClient, TasksDataClient>(config, client);
                 services.AddCompleteApiClient<ICsvExportClient, CsvExportClient>(config, client);
                 services.AddCompleteApiClient<IUsersClient, UsersClient>(config, client);
                 services.AddCompleteApiClient<IServiceSupportClient, ServiceSupportClient>(config, client);
@@ -124,7 +124,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 fixture.Inject(client);
                 fixture.Inject(serviceProvider.GetRequiredService<IProjectsClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IProjectGroupClient>());
-                fixture.Inject(serviceProvider.GetRequiredService<ITransferTasksDataClient>());
+                fixture.Inject(serviceProvider.GetRequiredService<ITasksDataClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<ICsvExportClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IUsersClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IServiceSupportClient>());