+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/user-capabilities/regional-caseworker-services/regional-casework-services-user.cy.ts
@@ -18,8 +18,7 @@ import projectRemover from "cypress/api/projectRemover";
 import projectApi from "cypress/api/projectApi";
 import { ProjectBuilder } from "cypress/api/projectBuilder";
 
-const date = new Date("2027-04-01");
-const project = ProjectBuilder.createConversionProjectRequest(date);
+const project = ProjectBuilder.createConversionProjectRequest({ significantDate: "2027-04-01" });
 let projectId: string;
 describe("Capabilities and permissions of the regional casework services user", () => {
     before(() => {