+++ CHANGELOG.md
@@ -17,14 +17,49 @@ Security in case of vulnerabilities.
 ## Unreleased
 
 ### Added
-- Added Project Significant History Dates `/projects/{projectId}/date-history`
-- Added `/projects/{projectId}/external-contacts` page
 
 ### Changed
 
 ### Fixed
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-05.608...main) for everything awaiting release
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-20.658...main) for everything awaiting release
+
+---
+
+## [1.13.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-08-20.658) - 2025-08-20
+
+### Added
+- Added `Conversion` and `Transfer` task list with their statuses.
+- Project groups: new API endpoints
+  - `/v1/ProjectGroup/List` – list project groups (with establishments)
+  - `/v1/ProjectGroup/Details` – project group details by id
+- Frontend routes and pages for project groups
+    - `/groups` – project groups list
+    - `/groups/{groupId}` – project group details
+- Added `Handover with Regional Delivery Officer` task page
+
+### Changed
+
+### Fixed
+- local authority name missing on delete local authority dialog
+- add missing incoming trust details on project page
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-12.625...production-2025-08-20.658) for everything in the release
+
+---
+
+## [1.12.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-08-12.625) - 2025-08-12
+
+### Added
+- Added Project Significant History Dates `/projects/{projectId}/date-history`
+- Added `/projects/{projectId}/external-contacts` page
+- Added `/privacy` page
+
+### Changed
+
+### Fixed
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-05.608...production-2025-08-12.625) for everything in the release
 
 ---
 
@@ -77,6 +112,10 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ### Added
 - New route `/projects/all/reports`
 - Add reports landing page under All projects > Reports (`/projects/all/reports`)
+- New route `/projects/{projectId}/date-history`
+- New route `/projects/{projectId}/date-history/new`
+- New route `/projects/{projectId}/date-history/reason`
+- New route `/projects/{projectId}/date-history/reasons/later`
 
 ### Changed  
 - Add a redirect from projects/{id} to projects/{id}/tasks